pkg/archive: move DecompressStream, CompressStream to compression pkg
