+++ pkg/archive/compression/compression_fuzz_test.go
@@ -1,0 +1,13 @@
+package compression
+
+import (
+	"bytes"
+	"testing"
+)
+
+func FuzzDecompressStream(f *testing.F) {
+	f.Fuzz(func(t *testing.T, data []byte) {
+		r := bytes.NewReader(data)
+		_, _ = DecompressStream(r)
+	})
+}
